week1: Attempt at karatsuba
It works in some cases but not others,
  also will not work for the 64 bit int case

Going to pivot and try a diff solution
