+++ Week 1/main.go
@@ -1,0 +1,10 @@
+package main
+
+import "fmt"
+
+func main() {
+	fmt.Println(recursiveIntegerMutliplication(1234, 5678))
+	fmt.Println(karatsuba(1234, 5678))
+	//fmt.Println(karatsuba(123, 345))
+	//fmt.Println(karatsuba(4444, 2222))
+}
